feat: add environment validation and update SupabasePlatformAdapter
- Introduced runtime environment validation for required Supabase variables
- Added assertSupabaseEnv function to ensure all necessary env vars are present
- Updated SupabasePlatformAdapter to validate env variables at startup
- Removed deprecated env-validation.ts file
- Enhanced type safety for environment variables in platform adapter
